Add exception handling

+++ userinfo.py
@@ -21,14 +21,20 @@ class Myinfo(Resource):
                 'status' : 400
             }
 
-        name, tel = self.getUserInfo()
+        success, name, tel = self.getUserInfo()
 
-        return {
-            'name' : name,
-            'tel' : tel,
-            'message' : 'success',
-            'status' : 200
-        }
+        if success:
+            return {
+                'name' : name,
+                'tel' : tel,
+                'message' : 'success',
+                'status' : 200
+            }
+        else:
+            return {
+                'message' : 'Invalid user info',
+                'status' : 40000
+            }
         
 
     def delete(self):
@@ -42,12 +48,19 @@ class Myinfo(Resource):
                 'status' : 400
             }
 
-        self.deleteUser()
+        success = self.deleteUser()
 
-        return {
-            'message' : 'user delete success',
-            'status' : 200
-        }
+        if success:
+            return {
+                'message' : 'user delete success',
+                'status' : 200
+            }
+        else:
+            return {
+                'message' : 'Invalid user info',
+                'status' : 40000
+            }
+        
 
 
     def put(self):
@@ -63,8 +76,9 @@ class Myinfo(Resource):
                 'status' : 400
             }
 
-        result = self.changePassword()
-        if result:
+        success = self.changePassword()
+
+        if success:
             return {
                 'message' : 'change password success',
                 'status' : 200
@@ -81,19 +95,35 @@ class Myinfo(Resource):
         query = f'SELECT name, tel FROM Profiles WHERE email="{self.email}"'
 
         cursor.execute(query)
-        name, tel = cursor.fetchall()[0]
+        result = cursor.fetchall()
+        if len(result) == 0:
+            return False, None, None
+        
+        name, tel = result[0]
 
         disconnect_database(db)
 
-        return name, tel
+        return True, name, tel
 
     
     def deleteUser(self):
         db, cursor = connect_database()
+        query = f'SELECT 1 from Profiles WHERE email="{self.email}"'
+        cursor.execute(query)
+
+        is_valid = len(cursor.fetchall()) != 0
+
+        if not is_valid:
+            return False
+
         query = f'DELETE FROM Profiles WHERE email="{self.email}"'
 
         cursor.execute(query)
         db.commit()
+
+        disconnect_database(db)
+
+        return True
 
     
     def changePassword(self):
@@ -110,5 +140,7 @@ class Myinfo(Resource):
 
         cursor.execute(query)
         db.commit()
+
+        disconnect_database(db)
 
         return True
